Switching to ESLint from JSHint + JSCS.
ESLint is a single, actively-maintained linter, which is easy to
configure and has good presets. Switching to it allowed me to
identify and correct a number of existing code quality issues.

A lot of the work in this PR will probably be overriden by the move
to the v2 component model later, but it's best to do things one
step at a time :)

Some highlights:
- Splits 'lint' task in two: one for sources, one for aux files.
- Sets two types of linting failures: errors and warnings.
  The tasks only fail for errors.
- Adjusts line lengths all over the place (sticking to 80 char).
- Doesn't switch any of the source to ES6, since we don't have the
  plumbing in place for that yet.